showenseignants.jsp, showetudiants.jsp, showprofileca.jsp, showprofileetudiant.jsp, showprofileenseignant.jsp
Created the showenseignants.jsp file, which is the list of all enseignants. Created the showetudiants.jsp which is the list of all etudiants. Created the showprofile.jsp which the profile of th CA. Created the showprofileetudiant.jsp which is the profile of the etudiant. Created the showenseignant.jsp which is the profile of the enseignant.